Workaround for part of initial login prompt arriving late.
If the initial 220 login prompt messages are multiple lines and
slow in arriving, we may interpret them as responses to the
initial login, and thus fail.  i.e. the following may occur:

220 Please log in.  More info:

<We send the username here>

220 This server is very interesting!
331 Anonymous login OK, please send password.

Here the library will interpret the second 220 status as the
response to the username, and will miss the actual 331 response.

The workaround is to discard any lines starting with "22" status
codes after sending the username.